Made changes?

+++ flaskr/fixation/tests/test_export.py
@@ -1,7 +1,7 @@
 import pytest
 from fixation_packages.event import Event
 from fixation_packages.event_list import EventList
-from fixation_packages.export import create_timestamp_list, create_json, write_str_to_file
+from fixation_packages.export import create_timestamp_list, create_json, write_json_to_file
 import numpy as np
 import json
 
@@ -16,12 +16,13 @@ def setup():
 class TestExport:
     def test_generate_timestamp_list(self, setup):
         out_list = create_timestamp_list(setup)
-        agh = np.array([(0, 1), (2.0, 2.5)])
+        agh = np.array([(np.float64(0), np.float64(1)), (np.float64(2.0), np.float64(2.5))])
         assert (out_list==agh).all()
 
     def test_generate_json_valid(self):
         test = "[[0.0, 1.0], [2.0, 2.5]]"
         out = create_json(np.array([(0, 1), (2.0, 2.5)]).tolist())
+        # write_json_to_file(out, "export.json")
         assert test == out
 
     def test_generate_json_no_input(setup):
